requirements.txt
h5py
plotly
matplotlib
opencv-python
numpy
python-dotenv
pandas